Rigid body to develop

+++ MathPhysicEngine/Rigidbody/Quaternion.cpp
@@ -1,10 +1,11 @@
 #include "Quaternion.h"
 #include "../Vecteur3D.h"
 #include <math.h>
+#include <iostream>
 
 void Quaternion::Normalized()
 {
-	float mag = 0;
+	double mag = 0;
 	for (int i = 0; i < 4; i++) {
 		mag += value[i] * value[i];
 	}
@@ -15,7 +16,7 @@ void Quaternion::Normalized()
 		return;
 	}
 
-	mag = (1.0f) / sqrtf(mag);
+	mag = (1.0f) / sqrt(mag);
 
 	for (int i = 0; i < 4; i++) {
 		value[i] *= mag;
@@ -72,12 +73,13 @@ void Quaternion::RotateByVector(const Vector3D& vector)
 
 }
 
-void Quaternion::UpdateByAngularVelocity(const Vector3D& rotation, float duration)
+void Quaternion::UpdateByAngularVelocity(const Vector3D& rotation, double duration)
 {
 	Quaternion q(0,
 		rotation.x * duration,
 		rotation.y * duration,
 		rotation.z * duration);
+
 
 	q *= *this;
 
@@ -85,4 +87,5 @@ void Quaternion::UpdateByAngularVelocity(const Vector3D& rotation, float duratio
 	value[1] += q.value[1] * (0.5f);
 	value[2] += q.value[2] * (0.5f);
 	value[3] += q.value[3] * (0.5f);
+
 }
