+++ MathPhysicEngine/Vecteur3D.h
@@ -6,35 +6,35 @@ class Vector3D
 
 public:
 
-	Vector3D(float x_ = 0, float y_ = 0, float z_ = 0);
+	Vector3D(double x_ = 0, double y_ = 0, double z_ = 0);
 	Vector3D(const Vector3D& v);
 
 	Vector3D operator+(const Vector3D& other) const;
 	Vector3D operator+=(const Vector3D& other);
 	Vector3D operator-(const Vector3D& other) const;
 	Vector3D operator-=(const Vector3D& other);
-	Vector3D operator*(float scalar) const;
-	Vector3D operator*=(float scalar);
-	Vector3D operator/(float scalar) const;
-	Vector3D operator/=(float scalar);
+	Vector3D operator*(double scalar) const;
+	Vector3D operator*=(double scalar);
+	Vector3D operator/(double scalar) const;
+	Vector3D operator/=(double scalar);
 	Vector3D& operator=(const Vector3D& other);
 
 	bool operator==(const Vector3D& other) const;
 
 	void Display();
 
-	bool Equal(const Vector3D& other, float epsilon = 1e-3) const;
+	bool Equal(const Vector3D& other, double epsilon = 1e-3) const;
 
-	float DotProduct(const Vector3D& other);
+	double DotProduct(const Vector3D& other);
 
 	Vector3D CrossProduct(const Vector3D& other);
 
 	//return magnitude/length of a vector
-	float GetNorm() const; 
+	double GetNorm() const; 
 
 	Vector3D Normalize();
 
-	float DistanceTo(const Vector3D& other);
+	double DistanceTo(const Vector3D& other);
 
 
 	friend std::ostream& operator<<(std::ostream& os, const Vector3D& v) {
@@ -42,11 +42,11 @@ public:
 		return os;
 	}
 
-	float x;
-	float y;
-	float z;
+	double x;
+	double y;
+	double z;
 
-	friend Vector3D operator*(float scalar, const Vector3D& vector);
+	friend Vector3D operator*(double scalar, const Vector3D& vector);
 
 };
